Addressing reviewer's comments
