+++ src/main/java/edu/ie3/datamodel/io/naming/DatabaseNamingStrategy.java
@@ -12,10 +12,21 @@ public class DatabaseNamingStrategy {
 
   private static final String TIME_SERIES_PREFIX = "time_series_";
 
+  /**
+   * Provides the String that all time series tables are prefixed with
+   *
+   * @return the time series prefix
+   */
   public String getTimeSeriesPrefix() {
     return TIME_SERIES_PREFIX;
   }
 
+  /**
+   * Provides the name of a time series table given a column scheme
+   *
+   * @param columnScheme the column scheme of the source data
+   * @return the table name
+   */
   public String getTimeSeriesEntityName(ColumnScheme columnScheme) {
     return TIME_SERIES_PREFIX + columnScheme.getScheme();
   }
