+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvTimeSeriesSource.java
@@ -10,11 +10,10 @@ import edu.ie3.datamodel.io.csv.CsvIndividualTimeSeriesMetaInformation;
 import edu.ie3.datamodel.io.factory.timeseries.*;
 import edu.ie3.datamodel.io.naming.FileNamingStrategy;
 import edu.ie3.datamodel.io.source.TimeSeriesSource;
-import edu.ie3.datamodel.io.source.TimeSeriesUtils;
 import edu.ie3.datamodel.models.timeseries.individual.IndividualTimeSeries;
 import edu.ie3.datamodel.models.timeseries.individual.TimeBasedValue;
 import edu.ie3.datamodel.models.value.*;
-import edu.ie3.datamodel.utils.TimeSeriesUtil;
+import edu.ie3.datamodel.utils.TimeSeriesUtils;
 import edu.ie3.util.interval.ClosedInterval;
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
@@ -165,7 +164,7 @@ public class CsvTimeSeriesSource<V extends Value> extends CsvDataSource
 
   @Override
   public IndividualTimeSeries<V> getTimeSeries(ClosedInterval<ZonedDateTime> timeInterval) {
-    return TimeSeriesUtil.trimTimeSeriesToInterval(timeSeries, timeInterval);
+    return TimeSeriesUtils.trimTimeSeriesToInterval(timeSeries, timeInterval);
   }
 
   @Override
