+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvTimeSeriesTypeSource.java
@@ -8,14 +8,24 @@ package edu.ie3.datamodel.io.source.csv;
 import edu.ie3.datamodel.io.naming.FileNamingStrategy;
 import edu.ie3.datamodel.io.naming.timeseries.IndividualTimeSeriesMetaInformation;
 import edu.ie3.datamodel.io.source.TimeSeriesTypeSource;
-import edu.ie3.datamodel.io.source.TimeSeriesUtils;
+import edu.ie3.datamodel.utils.TimeSeriesUtils;
 import java.util.Map;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+/**
+ * CSV implementation for retrieving {@link TimeSeriesTypeSource} from input directory structures
+ */
 public class CsvTimeSeriesTypeSource extends CsvDataSource implements TimeSeriesTypeSource {
 
-  protected CsvTimeSeriesTypeSource(
+  /**
+   * Creates a time series type source
+   *
+   * @param csvSep the CSV separator
+   * @param folderPath path that time series reside in
+   * @param fileNamingStrategy the file naming strategy
+   */
+  public CsvTimeSeriesTypeSource(
       String csvSep, String folderPath, FileNamingStrategy fileNamingStrategy) {
     super(csvSep, folderPath, fileNamingStrategy);
   }
