+++ src/main/java/edu/ie3/datamodel/io/source/sql/SqlTimeSeriesSource.java
@@ -13,10 +13,10 @@ import edu.ie3.datamodel.io.naming.DatabaseNamingStrategy;
 import edu.ie3.datamodel.io.naming.timeseries.ColumnScheme;
 import edu.ie3.datamodel.io.naming.timeseries.IndividualTimeSeriesMetaInformation;
 import edu.ie3.datamodel.io.source.TimeSeriesSource;
-import edu.ie3.datamodel.io.source.TimeSeriesUtils;
 import edu.ie3.datamodel.models.timeseries.individual.IndividualTimeSeries;
 import edu.ie3.datamodel.models.timeseries.individual.TimeBasedValue;
 import edu.ie3.datamodel.models.value.Value;
+import edu.ie3.datamodel.utils.TimeSeriesUtils;
 import edu.ie3.util.interval.ClosedInterval;
 import java.sql.Timestamp;
 import java.time.ZonedDateTime;
@@ -25,6 +25,7 @@ import java.util.*;
 public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBasedValue<V>>
     implements TimeSeriesSource<V> {
   private static final String WHERE = " WHERE ";
+  private static final String TIME_SERIES = "time_series";
 
   private final UUID timeSeriesUuid;
   private final Class<V> valueClass;
@@ -38,38 +39,6 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
 
   private final String queryTimeInterval;
   private final String queryTime;
-
-  /**
-   * Factory method to build a source from given meta information
-   *
-   * @param connector the connector needed for database connection
-   * @param schemaName the database schema to use
-   * @param tableName the database table to use
-   * @param metaInformation the time series meta information
-   * @param timePattern the pattern of time values
-   * @return a SqlTimeSeriesSource for given time series table
-   * @throws SourceException if the column scheme is not supported
-   * @deprecated since 3.0. Use {@link #createSource(SqlConnector, String, DatabaseNamingStrategy,
-   *     edu.ie3.datamodel.io.naming.timeseries.IndividualTimeSeriesMetaInformation, String)}
-   *     instead.
-   */
-  @Deprecated(since = "3.0", forRemoval = true)
-  public static SqlTimeSeriesSource<? extends Value> getSource(
-      SqlConnector connector,
-      String schemaName,
-      String tableName,
-      edu.ie3.datamodel.io.csv.timeseries.IndividualTimeSeriesMetaInformation metaInformation,
-      String timePattern)
-      throws SourceException {
-    if (!TimeSeriesSource.isSchemeAccepted(metaInformation.getColumnScheme()))
-      throw new SourceException(
-          "Unsupported column scheme '" + metaInformation.getColumnScheme() + "'.");
-
-    Class<? extends Value> valClass = metaInformation.getColumnScheme().getValueClass();
-
-    return create(
-        connector, schemaName, tableName, metaInformation.getUuid(), valClass, timePattern);
-  }
 
   /**
    * Factory method to build a source from given meta information
@@ -99,24 +68,6 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
         connector, schemaName, namingStrategy, metaInformation.getUuid(), valClass, timePattern);
   }
 
-  /**
-   * @deprecated since 3.0. Use {@link #create(SqlConnector, String, DatabaseNamingStrategy, UUID,
-   *     Class, String)} instead
-   */
-  @Deprecated(since = "3.0", forRemoval = true)
-  private static <T extends Value> SqlTimeSeriesSource<T> create(
-      SqlConnector connector,
-      String schemaName,
-      String tableName,
-      UUID timeSeriesUuid,
-      Class<T> valClass,
-      String timePattern) {
-    TimeBasedSimpleValueFactory<T> valueFactory =
-        new TimeBasedSimpleValueFactory<>(valClass, timePattern);
-    return new SqlTimeSeriesSource<>(
-        connector, schemaName, tableName, timeSeriesUuid, valClass, valueFactory);
-  }
-
   private static <T extends Value> SqlTimeSeriesSource<T> create(
       SqlConnector connector,
       String schemaName,
@@ -128,39 +79,6 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
         new TimeBasedSimpleValueFactory<>(valClass, timePattern);
     return new SqlTimeSeriesSource<>(
         connector, schemaName, namingStrategy, timeSeriesUuid, valClass, valueFactory);
-  }
-
-  /**
-   * Initializes a new SqlTimeSeriesSource
-   *
-   * @param connector the connector needed for database connection
-   * @param schemaName the database schema to use
-   * @param tableName the database table to use
-   * @param timeSeriesUuid the uuid of the time series
-   * @param valueClass the class of returned time series values
-   * @param factory a factory that parses the input data
-   * @deprecated since 3.0. Use {@link #SqlTimeSeriesSource(SqlConnector,
-   *     String,DatabaseNamingStrategy, UUID, Class, TimeBasedSimpleValueFactory)} instead
-   */
-  @Deprecated(since = "3.0", forRemoval = true)
-  public SqlTimeSeriesSource(
-      SqlConnector connector,
-      String schemaName,
-      String tableName,
-      UUID timeSeriesUuid,
-      Class<V> valueClass,
-      TimeBasedSimpleValueFactory<V> factory) {
-    super(connector);
-    this.timeSeriesUuid = timeSeriesUuid;
-    this.valueClass = valueClass;
-    this.valueFactory = factory;
-
-    String dbTimeColumnName = getDbColumnName(factory.getTimeFieldString(), tableName);
-
-    this.queryFull = createBaseQueryString(schemaName, tableName);
-    this.queryTimeInterval =
-        createQueryStringForTimeInterval(schemaName, tableName, dbTimeColumnName);
-    this.queryTime = createQueryStringForTime(schemaName, tableName, dbTimeColumnName);
   }
 
   /**
@@ -247,28 +165,10 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
   private String createQueryFull(String schemaName, String tableName) {
     return createBaseQueryString(schemaName, tableName)
         + WHERE
-        + "time_series = '"
+        + TIME_SERIES
+        + " = '"
         + timeSeriesUuid.toString()
         + "'";
-  }
-
-  /**
-   * Creates a base query to retrieve all entities in the given time frame with the following
-   * pattern: <br>
-   * {@code <base query> WHERE <time column> BETWEEN ? AND ?;}
-   *
-   * @param schemaName the name of the database schema
-   * @param tableName the name of the database table
-   * @param timeColumnName the name of the column holding the timestamp info
-   * @return the query string
-   */
-  @Deprecated(since = "3.0", forRemoval = true)
-  private static String createQueryStringForTimeInterval(
-      String schemaName, String tableName, String timeColumnName) {
-    return createBaseQueryString(schemaName, tableName)
-        + WHERE
-        + timeColumnName
-        + " BETWEEN ? AND ?;";
   }
 
   /**
@@ -285,26 +185,12 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
       String schemaName, String tableName, String timeColumnName) {
     return createBaseQueryString(schemaName, tableName)
         + WHERE
-        + "time_series = '"
+        + TIME_SERIES
+        + " = '"
         + timeSeriesUuid.toString()
         + "' AND "
         + timeColumnName
         + " BETWEEN ? AND ?;";
-  }
-
-  /**
-   * Creates a basic query to retrieve an entry for the given time with the following pattern: <br>
-   * {@code <base query> WHERE <time column>=?;}
-   *
-   * @param schemaName the name of the database schema
-   * @param weatherTableName the name of the database table
-   * @param timeColumnName the name of the column holding the timestamp info
-   * @return the query string
-   */
-  @Deprecated(since = "3.0", forRemoval = true)
-  private String createQueryStringForTime(
-      String schemaName, String weatherTableName, String timeColumnName) {
-    return createBaseQueryString(schemaName, weatherTableName) + WHERE + timeColumnName + "=?;";
   }
 
   /**
@@ -320,7 +206,8 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
   private String createQueryForTime(String schemaName, String tableName, String timeColumnName) {
     return createBaseQueryString(schemaName, tableName)
         + WHERE
-        + "time_series = '"
+        + TIME_SERIES
+        + " = '"
         + timeSeriesUuid.toString()
         + "' AND "
         + timeColumnName
