+++ src/test/groovy/edu/ie3/datamodel/utils/TimeSeriesUtilsTest.groovy
@@ -12,13 +12,13 @@ import spock.lang.Specification
 import java.time.ZoneId
 import java.time.ZonedDateTime
 
-class TimeSeriesUtilTest extends Specification implements TimeSeriesTestData {
+class TimeSeriesUtilsTest extends Specification implements TimeSeriesTestData {
 	def "A time series util is able to trim an individual time series to a given interval"() {
 		given:
 		def interval = new ClosedInterval(ZonedDateTime.of(1990, 1, 1, 0, 15, 0, 0, ZoneId.of("UTC")), ZonedDateTime.of(1990, 1, 1, 0, 30, 0, 0, ZoneId.of("UTC")))
 
 		when:
-		def actual = TimeSeriesUtil.trimTimeSeriesToInterval(individualIntTimeSeries, interval)
+		def actual = TimeSeriesUtils.trimTimeSeriesToInterval(individualIntTimeSeries, interval)
 
 		then:
 		actual.entries.size() == 2
@@ -29,7 +29,7 @@ class TimeSeriesUtilTest extends Specification implements TimeSeriesTestData {
 		def interval = new ClosedInterval(ZonedDateTime.of(1990, 12, 1, 0, 15, 0, 0, ZoneId.of("UTC")), ZonedDateTime.of(1990, 12, 1, 0, 30, 0, 0, ZoneId.of("UTC")))
 
 		when:
-		def actual = TimeSeriesUtil.trimTimeSeriesToInterval(individualIntTimeSeries, interval)
+		def actual = TimeSeriesUtils.trimTimeSeriesToInterval(individualIntTimeSeries, interval)
 
 		then:
 		actual.entries.size() == 0
@@ -40,7 +40,7 @@ class TimeSeriesUtilTest extends Specification implements TimeSeriesTestData {
 		def interval = new ClosedInterval(ZonedDateTime.of(1990, 1, 1, 0, 15, 0, 0, ZoneId.of("UTC")), ZonedDateTime.of(1990, 1, 1, 1, 45, 0, 0, ZoneId.of("UTC")))
 
 		when:
-		def actual = TimeSeriesUtil.trimTimeSeriesToInterval(individualIntTimeSeries, interval)
+		def actual = TimeSeriesUtils.trimTimeSeriesToInterval(individualIntTimeSeries, interval)
 
 		then:
 		actual.entries.size() == 2
